+++ src/languages/django.ts
@@ -1,7 +1,6 @@
 import { embeddedIn } from '../shared/languages/templating';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
-import type { LanguageProto } from '../types';
+import type { Grammar, LanguageProto } from '../types';
 
 // Django/Jinja2 syntax definition for Prism.js <http://prismjs.com> syntax highlighter.
 // Mostly it works OK but can paint code incorrectly on complex html/template tag combinations.
@@ -48,6 +47,6 @@ export default {
 				'punctuation': /[{}[\](),.:;]/,
 			},
 		},
-		[tokenize]: embeddedIn('markup'),
+		$tokenize: embeddedIn('markup') as Grammar['$tokenize'],
 	},
 } as LanguageProto<'django'>;
