+++ src/languages/handlebars.ts
@@ -1,7 +1,6 @@
 import { embeddedIn } from '../shared/languages/templating';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
-import type { LanguageProto } from '../types';
+import type { Grammar, LanguageProto } from '../types';
 
 export default {
 	id: 'handlebars',
@@ -35,6 +34,6 @@ export default {
 				'variable': /[^!"#%&'()*+,\/;<=>@\[\\\]^`{|}~\s]+/,
 			},
 		},
-		[tokenize]: embeddedIn('markup'),
+		$tokenize: embeddedIn('markup') as Grammar['$tokenize'],
 	},
 } as LanguageProto<'handlebars'>;
