+++ src/languages/latte.ts
@@ -1,6 +1,5 @@
 import { insertBefore } from '../shared/language-util';
 import { embeddedIn } from '../shared/languages/templating';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
 import php from './php';
 import type { Grammar, GrammarToken, LanguageProto } from '../types';
@@ -67,7 +66,7 @@ export default {
 					},
 				},
 			},
-			[tokenize]: embeddedIn(markupLatte),
+			$tokenize: embeddedIn(markupLatte) as Grammar['$tokenize'],
 		};
 	},
 } as LanguageProto<'latte'>;
