+++ src/languages/smarty.ts
@@ -1,7 +1,6 @@
 import { embeddedIn } from '../shared/languages/templating';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
-import type { GrammarToken, LanguageProto } from '../types';
+import type { Grammar, GrammarToken, LanguageProto } from '../types';
 
 export default {
 	id: 'smarty',
@@ -106,7 +105,7 @@ export default {
 				greedy: true,
 				inside: smarty,
 			},
-			[tokenize]: embeddedIn('markup'),
+			$tokenize: embeddedIn('markup') as Grammar['$tokenize'],
 		};
 	},
 } as LanguageProto<'smarty'>;
