+++ src/languages/twig.ts
@@ -1,7 +1,6 @@
 import { embeddedIn } from '../shared/languages/templating';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
-import type { LanguageProto } from '../types';
+import type { Grammar, LanguageProto } from '../types';
 
 export default {
 	id: 'twig',
@@ -46,6 +45,6 @@ export default {
 				'punctuation': /[()\[\]{}:.,]/,
 			},
 		},
-		[tokenize]: embeddedIn('markup'),
+		$tokenize: embeddedIn('markup') as Grammar['$tokenize'],
 	},
 } as LanguageProto<'twig'>;
