+++ src/languages/xquery.ts
@@ -1,6 +1,5 @@
 import { getTextContent, Token } from '../core/classes/token';
 import { withoutTokenize } from '../shared/language-util';
-import { tokenize } from '../shared/symbols';
 import markup from './markup';
 import type { TokenStream } from '../core/classes/token';
 import type { Grammar, GrammarToken, LanguageProto } from '../types';
@@ -184,7 +183,7 @@ export default {
 			inside: 'xquery',
 		};
 
-		xquery[tokenize] = (code, grammar, Prism) => {
+		xquery.$tokenize = (code, grammar, Prism) => {
 			const tokens = Prism.tokenize(code, withoutTokenize(grammar));
 			walkTokens(tokens);
 			return tokens;
