+++ src/shared.ts
@@ -6,5 +6,5 @@
 
 export { getLanguage, setLanguage, isActive } from './shared/dom-util';
 export { extend, insertBefore } from './shared/language-util';
-export { rest, tokenize } from './shared/symbols';
+export { rest } from './shared/symbols';
 export { templating, embeddedIn } from './shared/languages/templating';
