+++ src/shared/symbols.ts
@@ -2,8 +2,3 @@
  * The symbol grammars use to indicate a rest grammar.
  */
 export const rest = Symbol.for('Prism rest');
-
-/**
- * The symbol grammars use for a custom tokenizer.
- */
-export const tokenize = Symbol.for('Prism tokenize');
